travel-be/requirement.txt
asgiref==3.2.3
Django==2.0
djangorestframework==3.10.3
greenlet==0.4.15
msgpack==0.6.2
psycopg2-binary==2.8.4
pynvim==0.4.0
pytz==2019.3
sqlparse==0.3.0
PyJWT==1.7.1
django-rest-swagger==2.2.0